# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NexaTalent Enterprise Recruitment API",
    "version": "2.0.0",
    "description": "All 5 spec phases. Real JWT (15m access / 7d refresh rotation), bcrypt, 20 roles, 17 permissions, file-persisted DB, zod validation, idempotent finance, interview-gated chat. Legacy paths kept for old frontend builds."
  },
  "servers": [{ "url": "http://localhost:5000" }],
  "security": [{ "bearerAuth": [] }],
  "components": { "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } } },
  "paths": {
...
    "/api/v1/leads/{id}/merge": { "post": { "summary": "Merge duplicates (moves links, trails history)", "responses": { "200": { "description": "ok" }, "422": { "description": "protected" } } } },
...
    "/api/v1/exports/log": { "post": { "summary": "Audit-log a CSV export", "responses": { "201": { "description": "logged" } } } },    "/api/v1/leads/{id}": { "put": { "summary": "Edit lead (audited)", "responses": { "200": { "description": "ok" } } }, "delete": { "summary": "Delete lead (Won blocked)", "responses": { "200": { "description": "ok" }, "422": { "description": "audit-protected" } } } },
...
  }
}